Moved all rules to one file
